scripts/requirements.txt
bs4
progressbar2
requests
pillow